+++ scrapy/middleware.py
@@ -8,7 +8,8 @@ from collections import defaultdict, deque
 from typing import TYPE_CHECKING, Any, TypeVar, cast
 
 from scrapy.exceptions import NotConfigured, ScrapyDeprecationWarning
-from scrapy.utils.defer import _process_chain, process_parallel
+from scrapy.utils.defer import ensure_awaitable
+from scrapy.utils.deprecate import argument_is_required
 from scrapy.utils.misc import build_from_crawler, load_object
 from scrapy.utils.python import global_object_name
 
@@ -31,7 +32,6 @@ if TYPE_CHECKING:
 logger = logging.getLogger(__name__)
 
 _T = TypeVar("_T")
-_T2 = TypeVar("_T2")
 
 
 class MiddlewareManager(ABC):
@@ -56,6 +56,7 @@ class MiddlewareManager(ABC):
         self.methods: dict[str, deque[Callable | tuple[Callable, Callable] | None]] = (
             defaultdict(deque)
         )
+        self._mw_methods_requiring_spider: set[Callable] = set()
         for mw in middlewares:
             self._add_middleware(mw)
 
@@ -158,34 +159,53 @@ class MiddlewareManager(ABC):
         )
         return cls(*middlewares, crawler=crawler)
 
-    def _add_middleware(self, mw: Any) -> None:
-        if hasattr(mw, "open_spider"):
-            self.methods["open_spider"].append(mw.open_spider)
-        if hasattr(mw, "close_spider"):
-            self.methods["close_spider"].appendleft(mw.close_spider)
+    def _add_middleware(self, mw: Any) -> None:  # noqa: B027
+        pass
 
-    def _process_parallel(
-        self, methodname: str, obj: _T, *args: Any
-    ) -> Deferred[list[_T2]]:
-        methods = cast(
-            "Iterable[Callable[Concatenate[_T, _P], _T2]]", self.methods[methodname]
-        )
-        return process_parallel(methods, obj, *args)
+    def _check_mw_method_spider_arg(self, method: Callable) -> None:
+        if argument_is_required(method, "spider"):
+            warnings.warn(
+                f"{method.__qualname__}() requires a spider argument,"
+                f" this is deprecated and the argument will not be passed in future Scrapy versions."
+                f" If you need to access the spider instance you can save the crawler instance"
+                f" passed to from_crawler() and use its spider attribute.",
+                category=ScrapyDeprecationWarning,
+                stacklevel=2,
+            )
+            self._mw_methods_requiring_spider.add(method)
 
-    async def _process_chain(self, methodname: str, obj: _T, *args: Any) -> _T:
+    async def _process_chain(
+        self,
+        methodname: str,
+        obj: _T,
+        *args: Any,
+        add_spider: bool = False,
+        always_add_spider: bool = False,
+    ) -> _T:
         methods = cast(
             "Iterable[Callable[Concatenate[_T, _P], _T]]", self.methods[methodname]
         )
-        return await _process_chain(methods, obj, *args)
+        for method in methods:
+            if always_add_spider or (
+                add_spider and method in self._mw_methods_requiring_spider
+            ):
+                obj = await ensure_awaitable(method(obj, *(*args, self._spider)))
+            else:
+                obj = await ensure_awaitable(method(obj, *args))
+        return obj
 
-    def open_spider(self, spider: Spider | None = None) -> Deferred[list[None]]:
-        if spider:
-            self._warn_spider_arg("open_spider")
-            self._set_compat_spider(spider)
-        return self._process_parallel("open_spider", self._spider)
+    def open_spider(
+        self, spider: Spider | None = None
+    ) -> Deferred[list[None]]:  # pragma: no cover
+        raise NotImplementedError(
+            "MiddlewareManager.open_spider() is no longer implemented"
+            " and will be removed in a future Scrapy version."
+        )
 
-    def close_spider(self, spider: Spider | None = None) -> Deferred[list[None]]:
-        if spider:
-            self._warn_spider_arg("close_spider")
-            self._set_compat_spider(spider)
-        return self._process_parallel("close_spider", self._spider)
+    def close_spider(
+        self, spider: Spider | None = None
+    ) -> Deferred[list[None]]:  # pragma: no cover
+        raise NotImplementedError(
+            "MiddlewareManager.close_spider() is no longer implemented"
+            " and will be removed in a future Scrapy version."
+        )