+++ scrapy/pipelines/__init__.py
@@ -7,16 +7,20 @@ See documentation in docs/item-pipeline.rst
 from __future__ import annotations
 
 import warnings
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING, Any, Callable, cast
+
+from twisted.internet.defer import Deferred, DeferredList
 
 from scrapy.exceptions import ScrapyDeprecationWarning
 from scrapy.middleware import MiddlewareManager
 from scrapy.utils.conf import build_component_list
-from scrapy.utils.defer import deferred_from_coro
+from scrapy.utils.defer import deferred_from_coro, maybeDeferred_coro
 from scrapy.utils.python import global_object_name
 
 if TYPE_CHECKING:
-    from twisted.internet.defer import Deferred
+    from collections.abc import Iterable
+
+    from twisted.python.failure import Failure
 
     from scrapy import Spider
     from scrapy.settings import Settings
@@ -30,9 +34,15 @@ class ItemPipelineManager(MiddlewareManager):
         return build_component_list(settings.getwithbase("ITEM_PIPELINES"))
 
     def _add_middleware(self, pipe: Any) -> None:
-        super()._add_middleware(pipe)
+        if hasattr(pipe, "open_spider"):
+            self.methods["open_spider"].append(pipe.open_spider)
+            self._check_mw_method_spider_arg(pipe.open_spider)
+        if hasattr(pipe, "close_spider"):
+            self.methods["close_spider"].appendleft(pipe.close_spider)
+            self._check_mw_method_spider_arg(pipe.close_spider)
         if hasattr(pipe, "process_item"):
             self.methods["process_item"].append(pipe.process_item)
+            self._check_mw_method_spider_arg(pipe.process_item)
 
     def process_item(self, item: Any, spider: Spider | None = None) -> Deferred[Any]:
         if spider:
@@ -45,4 +55,36 @@ class ItemPipelineManager(MiddlewareManager):
         return deferred_from_coro(self.process_item_async(item))
 
     async def process_item_async(self, item: Any) -> Any:
-        return await self._process_chain("process_item", item, self._spider)
+        return await self._process_chain("process_item", item, add_spider=True)
+
+    def _process_parallel(self, methodname: str) -> Deferred[list[None]]:
+        methods = cast("Iterable[Callable[..., None]]", self.methods[methodname])
+
+        def get_dfd(method: Callable[..., None]) -> Deferred[None]:
+            if method in self._mw_methods_requiring_spider:
+                return maybeDeferred_coro(method, self._spider)
+            return maybeDeferred_coro(method)
+
+        dfds = [get_dfd(m) for m in methods]
+        d: Deferred[list[tuple[bool, None]]] = DeferredList(
+            dfds, fireOnOneErrback=True, consumeErrors=True
+        )
+        d2: Deferred[list[None]] = d.addCallback(lambda r: [x[1] for x in r])
+
+        def eb(failure: Failure) -> Failure:
+            return failure.value.subFailure
+
+        d2.addErrback(eb)
+        return d2
+
+    def open_spider(self, spider: Spider | None = None) -> Deferred[list[None]]:
+        if spider:
+            self._warn_spider_arg("open_spider")
+            self._set_compat_spider(spider)
+        return self._process_parallel("open_spider")
+
+    def close_spider(self, spider: Spider | None = None) -> Deferred[list[None]]:
+        if spider:
+            self._warn_spider_arg("close_spider")
+            self._set_compat_spider(spider)
+        return self._process_parallel("close_spider")