+++ scrapy/spidermiddlewares/base.py
@@ -1,8 +1,10 @@
 from __future__ import annotations
 
+import warnings
 from typing import TYPE_CHECKING, Any
 
 from scrapy import Request, Spider
+from scrapy.exceptions import ScrapyDeprecationWarning
 
 if TYPE_CHECKING:
     from collections.abc import AsyncIterator, Iterable
@@ -53,15 +55,32 @@ class BaseSpiderMiddleware:
                 yield o
 
     def process_spider_output(
-        self, response: Response, result: Iterable[Any], spider: Spider
+        self, response: Response, result: Iterable[Any], spider: Spider | None = None
     ) -> Iterable[Any]:
+        if spider is not None:  # pragma: no cover
+            warnings.warn(
+                "Passing a spider argument to BaseSpiderMiddleware.process_spider_output()"
+                " is deprecated and the passed value is ignored.",
+                ScrapyDeprecationWarning,
+                stacklevel=2,
+            )
         for o in result:
             if (o := self._get_processed(o, response)) is not None:
                 yield o
 
     async def process_spider_output_async(
-        self, response: Response, result: AsyncIterator[Any], spider: Spider
+        self,
+        response: Response,
+        result: AsyncIterator[Any],
+        spider: Spider | None = None,
     ) -> AsyncIterator[Any]:
+        if spider is not None:  # pragma: no cover
+            warnings.warn(
+                "Passing a spider argument to BaseSpiderMiddleware.process_spider_output_async()"
+                " is deprecated and the passed value is ignored.",
+                ScrapyDeprecationWarning,
+                stacklevel=2,
+            )
         async for o in result:
             if (o := self._get_processed(o, response)) is not None:
                 yield o