+++ tests/test_pipelines.py
@@ -4,6 +4,7 @@ import pytest
 from twisted.internet.defer import Deferred, inlineCallbacks, succeed
 
 from scrapy import Request, Spider, signals
+from scrapy.crawler import Crawler
 from scrapy.exceptions import ScrapyDeprecationWarning
 from scrapy.pipelines import ItemPipelineManager
 from scrapy.utils.asyncio import call_later
@@ -19,9 +20,17 @@ from tests.mockserver.http import MockServer
 
 
 class SimplePipeline:
-    def process_item(self, item, spider):
+    def process_item(self, item):
         item["pipeline_passed"] = True
         return item
+
+
+class DeprecatedSpiderArgPipeline:
+    def open_spider(self, spider):
+        pass
+
+    def close_spider(self, spider):
+        pass
 
 
 class DeferredPipeline:
@@ -29,7 +38,7 @@ class DeferredPipeline:
         item["pipeline_passed"] = True
         return item
 
-    def process_item(self, item, spider):
+    def process_item(self, item):
         d = Deferred()
         d.addCallback(self.cb)
         d.callback(item)
@@ -37,7 +46,7 @@ class DeferredPipeline:
 
 
 class AsyncDefPipeline:
-    async def process_item(self, item, spider):
+    async def process_item(self, item):
         d = Deferred()
         call_later(0, d.callback, None)
         await maybe_deferred_to_future(d)
@@ -46,7 +55,7 @@ class AsyncDefPipeline:
 
 
 class AsyncDefAsyncioPipeline:
-    async def process_item(self, item, spider):
+    async def process_item(self, item):
         d = Deferred()
         loop = asyncio.get_event_loop()
         loop.call_later(0, d.callback, None)
@@ -57,7 +66,7 @@ class AsyncDefAsyncioPipeline:
 
 
 class AsyncDefNotAsyncioPipeline:
-    async def process_item(self, item, spider):
+    async def process_item(self, item):
         d1 = Deferred()
         from twisted.internet import reactor
 
@@ -184,6 +193,9 @@ class TestCustomPipelineManager:
             def open_spider(self, spider):  # pylint: disable=signature-differs
                 return super().open_spider(spider)
 
+            def close_spider(self, spider):  # pylint: disable=signature-differs
+                return super().close_spider(spider)
+
             def process_item(self, item, spider):  # pylint: disable=signature-differs
                 with pytest.warns(
                     ScrapyDeprecationWarning,
@@ -214,11 +226,19 @@ class TestCustomPipelineManager:
             ),
             pytest.warns(
                 ScrapyDeprecationWarning,
+                match=r"The close_spider\(\) method of .+\.CustomPipelineManager requires a spider argument",
+            ),
+            pytest.warns(
+                ScrapyDeprecationWarning,
                 match=r"The process_item\(\) method of .+\.CustomPipelineManager requires a spider argument",
             ),
             pytest.warns(
                 ScrapyDeprecationWarning,
                 match=r"Passing a spider argument to CustomPipelineManager.open_spider\(\) is deprecated",
+            ),
+            pytest.warns(
+                ScrapyDeprecationWarning,
+                match=r"Passing a spider argument to CustomPipelineManager.close_spider\(\) is deprecated",
             ),
             pytest.warns(
                 ScrapyDeprecationWarning,
@@ -254,7 +274,7 @@ class TestCustomPipelineManager:
 
             def process_item(self, item, spider):
                 for pipeline in self.pipelines:
-                    item = pipeline.process_item(item, spider)
+                    item = pipeline.process_item(item)
                 return succeed(item)
 
         items = []
@@ -294,3 +314,128 @@ class TestCustomPipelineManager:
             await maybe_deferred_to_future(crawler.crawl(mockserver=mockserver))
 
         assert len(items) == 1
+
+
+class TestMiddlewareManagerSpider:
+    """Tests for the deprecated spider arg handling in MiddlewareManager.
+
+    Here because MiddlewareManager doesn't have methods that could take a spider arg."""
+
+    @pytest.fixture
+    def crawler(self) -> Crawler:
+        return get_crawler(Spider)
+
+    def test_deprecated_spider_arg_no_crawler_spider(self, crawler: Crawler) -> None:
+        """Crawler is provided, but doesn't have a spider. The instance passed to the method is
+        ignored and raises a warning."""
+        mwman = ItemPipelineManager(crawler=crawler)
+        with (
+            pytest.warns(
+                ScrapyDeprecationWarning,
+                match=r"DeprecatedSpiderArgPipeline.open_spider\(\) requires a spider argument",
+            ),
+            pytest.warns(
+                ScrapyDeprecationWarning,
+                match=r"DeprecatedSpiderArgPipeline.close_spider\(\) requires a spider argument",
+            ),
+        ):
+            mwman._add_middleware(DeprecatedSpiderArgPipeline())
+        with (
+            pytest.warns(
+                ScrapyDeprecationWarning,
+                match=r"Passing a spider argument to ItemPipelineManager.open_spider\(\) is deprecated",
+            ),
+            pytest.raises(
+                ValueError,
+                match="ItemPipelineManager needs to access self.crawler.spider but it is None",
+            ),
+        ):
+            mwman.open_spider(DefaultSpider())
+        with (
+            pytest.warns(
+                ScrapyDeprecationWarning,
+                match=r"Passing a spider argument to ItemPipelineManager.close_spider\(\) is deprecated",
+            ),
+            pytest.raises(
+                ValueError,
+                match="ItemPipelineManager needs to access self.crawler.spider but it is None",
+            ),
+        ):
+            mwman.close_spider(DefaultSpider())
+
+    def test_deprecated_spider_arg_with_crawler(self, crawler: Crawler) -> None:
+        """Crawler is provided and has a spider, works. The instance passed to the method is ignored,
+        even if mismatched, but raises a warning."""
+        mwman = ItemPipelineManager(crawler=crawler)
+        crawler.spider = crawler._create_spider("foo")
+        with pytest.warns(
+            ScrapyDeprecationWarning,
+            match=r"Passing a spider argument to ItemPipelineManager.open_spider\(\) is deprecated",
+        ):
+            mwman.open_spider(DefaultSpider())
+        with pytest.warns(
+            ScrapyDeprecationWarning,
+            match=r"Passing a spider argument to ItemPipelineManager.close_spider\(\) is deprecated",
+        ):
+            mwman.close_spider(DefaultSpider())
+
+    def test_deprecated_spider_arg_without_crawler(self) -> None:
+        """The first instance passed to the method is used, with a warning. Mismatched ones raise an error."""
+        with pytest.warns(
+            ScrapyDeprecationWarning,
+            match="was called without the crawler argument",
+        ):
+            mwman = ItemPipelineManager()
+        with (
+            pytest.warns(
+                ScrapyDeprecationWarning,
+                match=r"DeprecatedSpiderArgPipeline.open_spider\(\) requires a spider argument",
+            ),
+            pytest.warns(
+                ScrapyDeprecationWarning,
+                match=r"DeprecatedSpiderArgPipeline.close_spider\(\) requires a spider argument",
+            ),
+        ):
+            mwman._add_middleware(DeprecatedSpiderArgPipeline())
+        with pytest.warns(
+            ScrapyDeprecationWarning,
+            match=r"Passing a spider argument to ItemPipelineManager.open_spider\(\) is deprecated",
+        ):
+            mwman.open_spider(DefaultSpider())
+        with (
+            pytest.warns(
+                ScrapyDeprecationWarning,
+                match=r"Passing a spider argument to ItemPipelineManager.close_spider\(\) is deprecated",
+            ),
+            pytest.raises(
+                RuntimeError, match="Different instances of Spider were passed"
+            ),
+        ):
+            mwman.close_spider(DefaultSpider())
+        mwman.close_spider()
+
+    def test_no_spider_arg_without_crawler(self) -> None:
+        """If no crawler and no spider arg, raise an error."""
+        with pytest.warns(
+            ScrapyDeprecationWarning,
+            match="was called without the crawler argument",
+        ):
+            mwman = ItemPipelineManager()
+        with (
+            pytest.warns(
+                ScrapyDeprecationWarning,
+                match=r"DeprecatedSpiderArgPipeline.open_spider\(\) requires a spider argument",
+            ),
+            pytest.warns(
+                ScrapyDeprecationWarning,
+                match=r"DeprecatedSpiderArgPipeline.close_spider\(\) requires a spider argument",
+            ),
+        ):
+            mwman._add_middleware(DeprecatedSpiderArgPipeline())
+        with (
+            pytest.raises(
+                ValueError,
+                match="has no known Spider instance",
+            ),
+        ):
+            mwman.open_spider()