+++ tests/test_spidermiddleware_depth.py
@@ -22,12 +22,6 @@ def crawler() -> Crawler:
 
 
 @pytest.fixture
-def spider(crawler: Crawler) -> Spider:
-    crawler.spider = crawler._create_spider("scrapytest.org")
-    return crawler.spider
-
-
-@pytest.fixture
 def stats(crawler: Crawler) -> Generator[StatsCollector]:
     assert crawler.stats is not None
     crawler.stats.open_spider()
@@ -42,15 +36,13 @@ def mw(crawler: Crawler) -> DepthMiddleware:
     return DepthMiddleware.from_crawler(crawler)
 
 
-def test_process_spider_output(
-    mw: DepthMiddleware, stats: StatsCollector, spider: Spider
-) -> None:
+def test_process_spider_output(mw: DepthMiddleware, stats: StatsCollector) -> None:
     req = Request("http://scrapytest.org")
     resp = Response("http://scrapytest.org")
     resp.request = req
     result = [Request("http://scrapytest.org")]
 
-    out = list(mw.process_spider_output(resp, result, spider))
+    out = list(mw.process_spider_output(resp, result))
     assert out == result
 
     rdc = stats.get_value("request_depth_count/1")
@@ -58,7 +50,7 @@ def test_process_spider_output(
 
     req.meta["depth"] = 1
 
-    out2 = list(mw.process_spider_output(resp, result, spider))
+    out2 = list(mw.process_spider_output(resp, result))
     assert not out2
 
     rdm = stats.get_value("request_depth_max")